requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo;python_version<"3.9"
black==23.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==3.2.22
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
gunicorn==21.2.0
importlib-metadata==6.7.0
importlib-resources==5.8.1
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pycodestyle==2.10.0
pyflakes==3.0.0
PyJWT==2.8.0
python-decouple==3.8
psycopg2-binary==2.9.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.8.9
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.5.0
